- Local game works - GameAutomaton is no longer a singleton - Added a HIG standard menu and toolbar (hum) - Loading of saved games still crashes
svn path=/trunk/playground/games/ksirk/; revision=660814

+++ ksirk/Sprites/animspritesgroup.cpp
@@ -77,7 +77,7 @@ void AnimSpritesGroup::oneArrived(AnimSprite* sprite)
   {
     sprite->arrival();
   }
-  if (m_numberArrived == size())
+  if (m_numberArrived == (unsigned int)size())
   {
     emit arrived(this);
     m_numberArrived = 0;
